requirements.txt
requests>=2.28.0
python-dotenv>=1.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
transformers>=4.40
torch>=2.0
datasets
sklearn-crfsuite
pytorch-crf
matplotlib
seaborn
joblib
numpy
scikit-learn